requirements.txt
torch
torchvision
torchaudio
numpy
scikit-learn
matplotlib
gensim
tqdm
pandas
PyExecJS